[bazel] Upgrade workspace to Bazel 2.0.0.
This change introduces the following improvements:

- Upgrade the workspace to support Bazel 2.0.0.
- Adds Travis CI tests for Bazel, xcodebuild, and pod lib lint.
- Standardizes the Bazel WORKSPACE, version, and flags configuration with Material iOS team conventions.
- Adds framework-style import support for Objective-C libraries that depend on this repo.
- Adds build tests.

In order to make `bazel test //...` work seamlessly from a local machine, this change unfortunately currently disables iOS 9 testing. This is because Catalina does not support iOS 9 testing and I was not able to find a way to conditionally include iOS 9 simulators as part of testing using `select` and Blaze platform constraints.